fix(application error): Fix client-side exception and improve stability
  This commit fixes a client-side exception by correcting the Web3 AppKit initialization and adding defensive coding to prevent
  crashes from undefined objects.
